Some minor George adjustments.
I've decided to do as my master. I will comment every method and class declaration with a header, to visually identify them easier. Also, from now on, I will implement the "onClick" method of the public interface "OnClickListener", and inside that, I will identify the view that was clicked with a "switch case" statement. As opposed to what everybody does, which is to create a new "OnClickListener" inside the "setOnClickListener" method and instantiate the "onClick" then.