Change of ".gitignore" plans.
I've just changed my mind regarding the ".gitignore" file. I've decided to include the ".metadata" folder of the Eclipse workspace in the repository, in order to include all of my settings and whatnot. The other way was fine too, because it stored everything else, and that let you import the project to any other workspace. But I figured, whoever wanted to do that, could also do it excluding the ".metadata" folder himself. So, what the hell, let's just put everything in. That way, when I go into any other Eclipse bundle (like my laptop, which is the main reason I'm doing this), I'll just have to select this workspace. And hopefully I'll even have my cool black theme set.