Revert ".gitignore" change of plans.
Okay, so our previous commit was a complete failure. It turns out that Eclipse is constantly using the ".metadata" folder to store logs, and other cached files, making our commits impossible to understand. I'll go back to the way it was, and maybe, from time to time, I'll be uploading the ".metadata" folder when switching computer.