Creation of a new activity to show the contents of a Post.
We've created a new "PostActivity" to show the contents of a Post, for which we had to modify the "PostList" model, the "PostListAdapter", the "PostListActivity", and, of course, the manifest. There's a bug that crashes the app when clicking on an item twice that says "The content of the adapter has changed but ListView did not receive a notification. Make sure the content of your adapter is not modified from a background thread, but only from the UI thread. Make sure your adapter calls notifyDataSetChanged() when its content changes.". We've also added some miner George changes.