Creation of a menu and a dialog window.
In the "PostActivity", we created a menu with two new items; one for sharing, and another one for adding to favorites, to practice implicit and explicit intents. The share button starts an implicit intent that the system will get and let you choose between a list of apps that accept its kind of type (in this case, "text/plain").  The favorites button will pass an explicit intent to the dialog window, and the activity will return a result of which button was pressed on it (accept or cancel). We've also added a field to our model to show the URL of a Post.